Move auditing to an aspect-based system.
The previous patch had a system where auditing happened on endpoints
only if they explicitly fired an AuditApplicationEvent. Obviously this
isn't super user-friendly, so this patch introduces an Aspect which
looks for an @Auditable annotation and automatically audits matching
endpoints. Additionally, this patch introduces a proper AuditCode type
which includes methods for automatically converting an audit begin code
to its equivalent Done or Failed versions.