Added a nice 404 handler that works with swagger-ui.
Spring Boot sets up a default ResourceHttpRequestHandler that acts as a
sink for all requests that aren't caught by any other handlers. It does
this because it thinks it's clever.
This resource handler prevents the firing of the HandlerNotFound
exception, and Boot offers no way to configure its erroring behaviour
that I could find, so to generate restful 404s we need to disable this
sink.
Most of the time this wouldn't matter for a restful api that doesn't
need to return static content (the only reason Boot does this). However,
we're using swagger-ui in this project, which *does* require static
content; mostly the swagger-ui.html and its sheets and JS. For this to
continue to function, we must add our own resource handler(s) that
specifically capture the swagger-ui.html path and the /webjars/** paths
that it uses to store its css and js and serve up the appropriate
resources.